requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
colorama>=0.4.6
pefile>=2023.2.7
cryptography>=41.0.0
py7zr>=0.20.6
python-magic>=0.4.27